docs/03-Tutorials/08-client-server-gpt-oss/pyproject.toml
[project]
name = "mcp-gpt-oss-tutorial"
version = "0.1.0"
description = "MCP Tutorial with Local GPT-OSS Integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "streamlit>=1.28.0",
    "ollama>=0.1.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "httpx",
    "sse-starlette",
    "uvicorn",
    "fastapi",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["libs"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
